增加库存管理

+++ factory-produce/src/main/java/com/app/controller/v1/repertory/GoodsAPI.java
@@ -14,8 +14,12 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.app.controller.common.Result;
 import com.app.dao.sql.SQLWhere;
+import com.app.dao.sql.cnd.INCnd;
 import com.app.dao.sql.cnd.LikeCnd;
+import com.app.dao.sql.cnd.NotEQCnd;
+import com.app.dao.sql.cnd.NotINCnd;
 import com.app.dao.sql.sort.DescSort;
+import com.app.entity.repertory.RepertoryGoodsComponentEntity;
 import com.app.entity.repertory.RepertoryGoodsEntity;
 import com.app.util.PublicMethod;
 import com.google.gson.JsonArray;
@@ -34,25 +38,24 @@ public class GoodsAPI extends Result{
     public static Log logger = LogFactory.getLog(GoodsAPI.class);
     
   
-    
+    /**
+     * 查询产品信息列表
+     */
     @RequestMapping(method = { RequestMethod.POST, RequestMethod.GET },value="/list")
     public String list(@RequestParam String aoData) throws Exception{
     	JsonArray jo = new JsonParser().parse(aoData).getAsJsonArray();
     	SQLWhere sql = new SQLWhere().orderBy(new DescSort(RepertoryGoodsEntity.GOODS_ID));
-    	int iDisplayStart = 0;// 起始  
-    	int iDisplayLength = 10;// size 
-    	int sEcho = 0;
     	for(JsonElement je : jo){
     		JsonObject jsonObject = je.getAsJsonObject();
-    		if (jsonObject.get(NAME).getAsString().equals(S_ECHO))  
-                sEcho = jsonObject.get(VALUE).getAsInt();  
+    		if (jsonObject.get(NAME).getAsString().equals(S_ECHO))
+                sEcho = jsonObject.get(VALUE).getAsInt();
             else if (jsonObject.get(NAME).getAsString().equals(I_DISPLAY_START))  
                 iDisplayStart = jsonObject.get(VALUE).getAsInt();  
             else if (jsonObject.get(NAME).getAsString().equals(I_DISPLAY_LENGTH))  
                 iDisplayLength = jsonObject.get(VALUE).getAsInt(); 
             else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.NAME)){  
             	sql.and(new LikeCnd(RepertoryGoodsEntity.NAME,jsonObject.get(VALUE).getAsString()));
-            }else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.CODE)){  
+            }else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.CODE)){
             	sql.and(new LikeCnd(RepertoryGoodsEntity.CODE,jsonObject.get(VALUE).getAsString()));
             }
     	}
@@ -61,6 +64,51 @@ public class GoodsAPI extends Result{
     	RepertoryGoodsEntity entity = new RepertoryGoodsEntity();
     	
     	
+    	List<RepertoryGoodsEntity> list = entity.getListVO(iDisplayStart, iDisplayLength, sql);
+    	
+    	
+    	long count = entity.getCount(sql);
+    	Map<String,Object> map = new HashMap<String,Object>();
+    	map.put("status", 200);
+    	map.put("data", list);
+    	map.put("sEcho", ++sEcho);
+    	map.put("iTotalRecords", count);
+    	map.put("iTotalDisplayRecords", count);
+        return success(map);
+    }
+    
+    
+    /**
+     * 获取非成品的产品数据
+     * @param aoData
+     * @return
+     * @throws Exception
+     */
+    @RequestMapping(method = { RequestMethod.POST, RequestMethod.GET },value="/no_finished_list")
+    public String noFinishedList(@RequestParam String aoData) throws Exception{
+    	JsonArray jo = new JsonParser().parse(aoData).getAsJsonArray();
+    	SQLWhere sql = new SQLWhere(new INCnd(RepertoryGoodsEntity.TYPE, new Integer[]{RepertoryGoodsEntity.GOOD_TYPE_MATERIAL,RepertoryGoodsEntity.GOOD_TYPE_SEMI_PRODUCT})).orderBy(new DescSort(RepertoryGoodsEntity.TYPE,RepertoryGoodsEntity.GOODS_ID));
+    	for(JsonElement je : jo){
+    		JsonObject jsonObject = je.getAsJsonObject();
+    		if (jsonObject.get(NAME).getAsString().equals(S_ECHO))
+                sEcho = jsonObject.get(VALUE).getAsInt();
+            else if (jsonObject.get(NAME).getAsString().equals(I_DISPLAY_START))  
+                iDisplayStart = jsonObject.get(VALUE).getAsInt();  
+            else if (jsonObject.get(NAME).getAsString().equals(I_DISPLAY_LENGTH))  
+                iDisplayLength = jsonObject.get(VALUE).getAsInt(); 
+            else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.NAME)){  
+            	sql.and(new LikeCnd(RepertoryGoodsEntity.NAME,jsonObject.get(VALUE).getAsString()));
+            }else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.CODE)){
+            	sql.and(new LikeCnd(RepertoryGoodsEntity.CODE,jsonObject.get(VALUE).getAsString()));
+            }else if (jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.CODE)){
+            	sql.and(new LikeCnd(RepertoryGoodsEntity.CODE,jsonObject.get(VALUE).getAsString()));
+            }else if(jsonObject.get(NAME).getAsString().equals(RepertoryGoodsEntity.GOODS_ID)){
+            	sql.and(new NotEQCnd(RepertoryGoodsEntity.GOODS_ID,jsonObject.get(VALUE).getAsLong())).and(new NotINCnd(RepertoryGoodsEntity.GOODS_ID,"select component_id from t_repertory_goods_component where goods_id="+jsonObject.get(VALUE).getAsLong()));
+            }
+    	}
+    	
+    	logger.error(aoData);
+    	RepertoryGoodsEntity entity = new RepertoryGoodsEntity();
     	List<RepertoryGoodsEntity> list = entity.getListVO(iDisplayStart, iDisplayLength, sql);
     	
     	
@@ -87,13 +135,31 @@ public class GoodsAPI extends Result{
         return success(entity);
     }
     
+    
+    
+    /**
+     * 删除
+     * @param id
+     * @return
+     */
     @RequestMapping(method=RequestMethod.DELETE,value="/delete/{id}")
     public String delete(@PathVariable("id") Long id) {
     	RepertoryGoodsEntity entity = new RepertoryGoodsEntity();
     	try{
+    		entity.setGoodsId(id).loadVo();
+    		if(entity.getInventory() > 0 || entity.getLocking() > 0){
+    			return error("库存大于0，不能为删除");
+    		}
+    		List<RepertoryGoodsComponentEntity>  list = new RepertoryGoodsComponentEntity().setGoodsId(id).queryCustomCacheValue(0, null);
+    		if(list != null && list.size() > 0){
+    			for(RepertoryGoodsComponentEntity gc : list){
+    				gc.delete();
+    			}
+    		}
     		entity.setGoodsId(id).delete();
     		return success("删除成功");
     	}catch(Exception e){
+    		logger.error("删除失败", e);
     		return success("删除失败");
     	}
     	
@@ -103,15 +169,31 @@ public class GoodsAPI extends Result{
     
     
     
-    
+    /**
+     * 修改产品信息
+     * 此时是否需要考滤将成品修改为原料时，
+     * 产品数据是否清除产品零件表的数据
+     * @param aoData
+     * @return
+     * @throws Exception
+     */
     @RequestMapping(method={ RequestMethod.POST, RequestMethod.PUT },value="/update")
     public String update(@RequestParam String aoData) throws Exception{
     	RepertoryGoodsEntity entity = new RepertoryGoodsEntity();
     	logger.error("-------"+aoData);
-    	entity.parse(new JsonParser().parse(aoData).getAsJsonObject());
+    	JsonObject jo = new JsonParser().parse(aoData).getAsJsonObject();
+    	jo.remove(RepertoryGoodsEntity.LOCKING);//去掉库存
+    	jo.remove(RepertoryGoodsEntity.INVENTORY);//去掉锁定
+    	jo.remove("inventory");//去掉库存
+    	jo.remove("locking");//去掉锁定
+    	entity.parse(jo);
     	
     	if(PublicMethod.isEmptyStr(entity.getName())){
-    		return error("人员名称不能为空");
+    		return error("产品名称不能为空");
+    	}
+    	
+    	if(PublicMethod.isEmptyStr(entity.getCode())){
+    		return error("产品编号不能为空");
     	}
     	
     	
@@ -120,13 +202,19 @@ public class GoodsAPI extends Result{
     		entity.update();
         	return success("修改成功",entity.getGoodsId());
     	}catch(Exception e){
+    		logger.error("修改失败", e);
     		return error(e.getMessage());
     	}
     	
     }
     
     
-    
+    /**
+     * 新增产品信息
+     * @param aoData
+     * @return
+     * @throws Exception
+     */
     @RequestMapping(method={ RequestMethod.POST, RequestMethod.PUT },value="/add")
     public String add(@RequestParam String aoData) throws Exception{
     	RepertoryGoodsEntity entity = new RepertoryGoodsEntity();
